+++ src/main/java/app/Bermuda.java
@@ -5,7 +5,7 @@ import app.permission.GetPermissionUseCaseFactory;
 import app.permission.UpdatePermissionUseCaseFactory;
 import app.project.CreateProjectUseCaseFactory;
 import app.project.GetProjectUseCaseFactory;
-import app.share_project.ShareProjectPageUseCaseFactory;
+import app.project.ShareProjectPageUseCaseFactory;
 import app.task.AddTaskUseCaseFactory;
 import app.task.EditTaskUseCaseFactory;
 import app.task.GetTaskUseCaseFactory;
@@ -19,8 +19,8 @@ import domains.permission.use_case.create_permission.CreatePermissionDataAccessI
 import domains.permission.use_case.delete_permission.DeletePermissionDataAccessInterface;
 import domains.permission.use_case.get_permission.GetPermissionDataAccessInterface;
 import domains.permission.use_case.update_permission.UpdatePermissionDataAccessInterface;
-import domains.project.share_project.ShareProjectDataAccessInterface;
-import domains.project.share_project_page.ShareProjectPageDataAccessInterface;
+import domains.project.use_case.share_project.ShareProjectDataAccessInterface;
+import domains.project.use_case.share_project_page.ShareProjectPageDataAccessInterface;
 import domains.project.use_case.create_project.CreateProjectApiDataAccessInterface;
 import domains.project.use_case.create_project.CreateProjectSqlDataAccessInterface;
 import domains.project.use_case.get_project.GetProjectApiDataAccessInterface;
