+++ src/main/java/data_access/cloudsql/SqlDataAccessObject.java
@@ -5,8 +5,8 @@ import domains.permission.use_case.create_permission.CreatePermissionDataAccessI
 import domains.permission.use_case.delete_permission.DeletePermissionDataAccessInterface;
 import domains.permission.use_case.get_permission.GetPermissionDataAccessInterface;
 import domains.permission.use_case.update_permission.UpdatePermissionDataAccessInterface;
-import domains.project.share_project.ShareProjectDataAccessInterface;
-import domains.project.share_project_page.ShareProjectPageDataAccessInterface;
+import domains.project.use_case.share_project.ShareProjectDataAccessInterface;
+import domains.project.use_case.share_project_page.ShareProjectPageDataAccessInterface;
 import domains.project.use_case.create_project.CreateProjectSqlDataAccessInterface;
 import domains.project.use_case.get_project.GetProjectSqlDataAccessInterface;
 import domains.user.entity.User;
