+++ src/main/java/domains/project/use_case/share_project/ShareProjectInputBoundary.java
@@ -1,4 +1,4 @@
-package domains.project.share_project;
+package domains.project.use_case.share_project;
 
 public interface ShareProjectInputBoundary {
     void execute(ShareProjectInputData input);
