+++ src/main/java/domains/project/use_case/share_project/ShareProjectInteractor.java
@@ -1,4 +1,4 @@
-package domains.project.share_project;
+package domains.project.use_case.share_project;
 
 import domains.permission.entity.NewPermissionFactory;
 public class ShareProjectInteractor implements ShareProjectInputBoundary {
