+++ src/main/java/domains/project/use_case/share_project_page/ShareProjectPageDataAccessInterface.java
@@ -1,4 +1,4 @@
-package domains.project.share_project_page;
+package domains.project.use_case.share_project_page;
 
 import java.util.List;
 public interface ShareProjectPageDataAccessInterface {
