+++ src/main/java/domains/project/use_case/share_project_page/ShareProjectPageOutputBoundary.java
@@ -1,4 +1,4 @@
-package domains.project.share_project_page;
+package domains.project.use_case.share_project_page;
 
 public interface ShareProjectPageOutputBoundary {
     void prepareSuccessView(ShareProjectPageOutputData output);
