+++ src/main/java/interface_adapter/project/share_project/ShareProjectController.java
@@ -1,7 +1,7 @@
 package interface_adapter.project.share_project;
 
-import domains.project.share_project.ShareProjectInputBoundary;
-import domains.project.share_project.ShareProjectInputData;
+import domains.project.use_case.share_project.ShareProjectInputBoundary;
+import domains.project.use_case.share_project.ShareProjectInputData;
 public class ShareProjectController {
     ShareProjectInputBoundary shareProjectInteractor;
     public ShareProjectController(
