+++ src/main/java/interface_adapter/project/share_project_page/ShareProjectPageController.java
@@ -1,7 +1,7 @@
 package interface_adapter.project.share_project_page;
 
-import domains.project.share_project_page.ShareProjectPageInputBoundary;
-import domains.project.share_project_page.ShareProjectPageInputData;
+import domains.project.use_case.share_project_page.ShareProjectPageInputBoundary;
+import domains.project.use_case.share_project_page.ShareProjectPageInputData;
 public class ShareProjectPageController {
     private final ShareProjectPageInputBoundary shareProjectInteractor;
